Consolidating new model into existing List processors.
