+++ nifi-nar-bundles/nifi-azure-bundle/nifi-azure-processors/src/main/java/org/apache/nifi/processors/azure/storage/utils/BlobInfo.java
@@ -209,4 +209,9 @@ public class BlobInfo implements Comparable<BlobInfo>, Serializable, ListableEnt
     public long getTimestamp() {
         return getLastModifiedTime();
     }
+
+    @Override
+    public long getSize() {
+        return length;
+    }
 }
