+++ nifi-nar-bundles/nifi-extension-utils/nifi-processor-utils/src/main/java/org/apache/nifi/processor/util/list/ListableEntity.java
@@ -37,4 +37,7 @@ public interface ListableEntity {
      */
     long getTimestamp();
 
+    // TODO: doc
+    long getSize();
+
 }
